feat(pay): three-rung transport ladder — AWDL, Nearby, QR floor
