+++ utils/localpay/transport/nearby.ts
@@ -1,0 +1,3 @@
+import { makeSocketTransport } from './socket'
+
+export const nearbyTransport = makeSocketTransport('nearby')
