+++ utils/localpay/transport/types.ts
@@ -61,7 +61,7 @@ export interface ReceivedFrame {
 }
 
 export interface LocalPaymentTransport {
-  readonly kind: 'awdl' | 'qr'
+  readonly kind: 'awdl' | 'nearby' | 'qr'
   receive(session: Session, signal: AbortSignal): Promise<ReceivedFrame>
   send(session: Session, frame: PaymentFrame, signal: AbortSignal): Promise<Ack>
 }
