+++ utils/pay/rails/nearby.ts
@@ -22,5 +22,12 @@ export {
 export { buildPaymentFrame, finalizeDelivery } from '@/utils/localpay/build'
 export { holdSentPaymentOffline } from '@/utils/offline/payerHold'
 export { awdlTransport } from '@/utils/localpay/transport/awdl'
-export { localSupportsAwdl, selectTransport } from '@/utils/localpay/transport/select'
+export { nearbyTransport } from '@/utils/localpay/transport/nearby'
+export {
+  localSupportsAwdl,
+  localSupportsNearby,
+  selectTransport,
+  type TransportKind
+} from '@/utils/localpay/transport/select'
 export { isDeclineReason, type Ack, type ConfirmDelivery, type DeclineReason } from '@/utils/localpay/transport/types'
+export { CAP_NEARBY } from '@/utils/localpay/session'
